Renamed linear_quadratic_regulator to lqr
