+++ include/robotics/linear_control/lqr.hpp
