+++ test/source/linear_control/lqr.cpp
@@ -1,7 +1,7 @@
 #include <doctest/doctest.h>
 
 #include <Eigen/Dense>
-#include <robotics/linear_control/linear_quadratic_regulator.hpp>
+#include <robotics/linear_control/lqr.hpp>
 #include <string>
 
 TEST_CASE("Robotics")
